refactor(st-09078): modularize streaming select executor
